split up tests into different files
